+++ nacc_attribute_deriver/attributes/nacc/modules/uds/form_d1.py
@@ -292,8 +292,8 @@ class UDSFormD1Attribute(AttributeCollection):
         naccnorm = self.__uds.get_cross_sectional_value("naccnorm")
 
         try:
-            if int(naccnorm) == 0:
-                return naccnorm
+            if naccnorm is not None and int(naccnorm) == 0:
+                return 0
 
             return int(self.__uds.get_value("normcog"))
         except (ValueError, TypeError):